Rename OpenIdServiceHttpError to its proper name
We should call the file the same as the error type.
