+++ src/core/adgangsplatformen/AdgangsPlatformenServiceHttpError.ts
